requirements.txt
requirements.txt

Pillow>=9.0.0

tkinterdnd2>=0.3.0

Como usar
No terminal:

pip install -r requirements.txt